Move collectionPage to common routes
