+++ packages/web/src/public-site/pages/landing-page/components/FeaturedContent.tsx
@@ -2,6 +2,7 @@ import { useEffect, useState } from 'react'
 
 import { useAppContext } from '@audius/common/context'
 import { SquareSizes, UserCollectionMetadata } from '@audius/common/models'
+import { collectionPage } from '@audius/common/src/utils/route'
 import { Nullable, Maybe } from '@audius/common/utils'
 import { StorageNodeSelectorService } from '@audius/sdk'
 // eslint-disable-next-line no-restricted-imports -- TODO: migrate to @react-spring/web
@@ -21,7 +22,6 @@ import useHasViewed from 'hooks/useHasViewed'
 import { handleClickRoute } from 'public-site/components/handleClickRoute'
 import { audiusBackendInstance } from 'services/audius-backend/audius-backend-instance'
 import { env } from 'services/env'
-import { collectionPage } from 'utils/route'
 
 import styles from './FeaturedContent.module.css'
